+++ packages/br-validators/src/ibpt.ts
@@ -1,0 +1,1 @@
+export * from './ibpt/index.js';